# bpf optimization syscall sampler